# frontend/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Inventory Management API
  version: 1.0.0
  description: API for managing product inventory and orders

servers:
  - url: /api

tags:
  - name: Products
    description: Product inventory management (producer)
  - name: Orders
    description: Order management (user + producer)

paths:
  /products:
    get:
      operationId: getProducts
      summary: List all products
      tags: [Products]
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: search
          in: query
          schema:
            type: string
          description: Search by product name
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      operationId: createProduct
      summary: Add a new product
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductInput'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/low-stock:
    get:
      operationId: getLowStockProducts
      summary: List products below their low-stock threshold
      tags: [Products]
      responses:
        '200':
          description: Low stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      operationId: getProduct
      summary: Get a single product by ID
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateProduct
      summary: Update product details
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductInput'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}/stock:
    patch:
      operationId: updateStock
      summary: Update product stock quantity
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockInput'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid stock update (e.g. would result in negative stock)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders:
    post:
      operationId: createOrder
      summary: Place a new order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderInput'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error (e.g. insufficient stock)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      operationId: getOrders
      summary: List all orders
      tags: [Orders]
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get a single order by ID
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      summary: Update order status
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusInput'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string

    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: string

  schemas:
    Product:
      type: object
      required: [_id, name, sku, description, category, quantity, price, lowStockThreshold, createdAt, updatedAt]
      properties:
        _id:
          type: string
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        category:
          type: string
        image:
          type: string
          format: uri
        expiryDate:
          type: string
          format: date-time
        quantity:
          type: integer
          minimum: 0
        price:
          type: number
          format: float
        lowStockThreshold:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateProductInput:
      type: object
      required: [name, sku, description, category, quantity, price]
      properties:
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        category:
          type: string
        image:
          type: string
          format: uri
        expiryDate:
          type: string
          format: date-time
        quantity:
          type: integer
          minimum: 0
        price:
          type: number
          format: float
        lowStockThreshold:
          type: integer
          minimum: 0
          default: 10

    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        image:
          type: string
          format: uri
        expiryDate:
          type: string
          format: date-time
        price:
          type: number
          format: float
        lowStockThreshold:
          type: integer
          minimum: 0

    UpdateStockInput:
      type: object
      required: [adjustment]
      properties:
        adjustment:
          type: integer
          description: Positive to add stock, negative to remove. Server rejects if result would be negative.

    OrderItem:
      type: object
      required: [product, quantity, priceAtOrder]
      properties:
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        priceAtOrder:
          type: number
          format: float

    Order:
      type: object
      required: [_id, items, status, totalAmount, createdAt, updatedAt]
      properties:
        _id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        totalAmount:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateOrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [product, quantity]
            properties:
              product:
                type: string
                description: Product ID
              quantity:
                type: integer
                minimum: 1

    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, cancelled]

    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string